move to VectorUtil
