+++ lucene/core/src/java/org/apache/lucene/internal/vectorization/DefaultVectorizationProvider.java
@@ -44,9 +44,4 @@ final class DefaultVectorizationProvider extends VectorizationProvider {
   public PostingDecodingUtil newPostingDecodingUtil(IndexInput input) {
     return new PostingDecodingUtil(input);
   }
-
-  @Override
-  public BitSetUtil newBitSetUtil() {
-    return BitSetUtil.INSTANCE;
-  }
 }
