+++ lucene/core/src/java/org/apache/lucene/internal/vectorization/VectorizationProvider.java
@@ -114,8 +114,6 @@ public abstract class VectorizationProvider {
   /** Create a new {@link PostingDecodingUtil} for the given {@link IndexInput}. */
   public abstract PostingDecodingUtil newPostingDecodingUtil(IndexInput input) throws IOException;
 
-  public abstract BitSetUtil newBitSetUtil();
-
   // *** Lookup mechanism: ***
 
   private static final Logger LOG = Logger.getLogger(VectorizationProvider.class.getName());
@@ -208,8 +206,7 @@ public abstract class VectorizationProvider {
           "org.apache.lucene.util.VectorUtil",
           "org.apache.lucene.codecs.lucene103.Lucene103PostingsReader",
           "org.apache.lucene.codecs.lucene103.PostingIndexInput",
-          "org.apache.lucene.tests.util.TestSysoutsLimits",
-          "org.apache.lucene.util.FixedBitSet");
+          "org.apache.lucene.tests.util.TestSysoutsLimits");
 
   private static void ensureCaller() {
     final boolean validCaller =
